moved components to user route
